Rewrite `Iter{,Mut}` to be a double-pointer implementation
This change reduces the computation done by `.next()` and
`.next_back()` by representing the iteration region as a pair of
boundary pointers.

The behavior is analagous to the double-pointer iteration seen in
`core::slice::Iter{,Mut}` and C++ iterators, so I will not explain
it further here. The implementation uses `BitIdx` rather than
`BitTail` for the ending marker, because the ending marker must be
considered a live pointer one past the end of the region, rather
than a dead pointer, in order to support `.next_back()` correctly.

Instruction size of `.next()` is considerably reduced, and the
runtime drops from 17ns to 8.

Benchmarks:

old:     2,634ns +/- 898
old_rev: 1,519ns +/- 256
new:     1,090ns +/- 246
new_rev: 1,095ns +/- 186